Move ConcurrentBitSet to separate class
